move setup files to the root javascript
